requirements.txt
thefirstock
pyotp
pandas
streamlit
numpy
faiss-cpu
requests
python-dotenv
Pillow
transformers
scikit-learn
ipywidgets
notebook
mplfinance
NorenRestApi-0.0.30-py2.py3-none-any.whl
Flask
fastapi
python-dotenv
uvicorn
